requirements.txt
# Core dependencies
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.25.0

# Data processing and utilities
numpy>=1.24.0
pandas>=2.0.0

# File handling
pathlib2>=2.3.7

# Optional: For additional file formats
# pypdf>=3.17.0  # For PDF processing
# faiss-cpu>=1.7.4  # For local vector search (if needed)
# tiktoken>=0.5.0  # For token counting 